requirements.txt
alembic==1.13.1
apifairy==1.4.0
apispec==6.7.0
blinker==1.8.2
click==8.1.7
Flask==3.0.3
Flask-HTTPAuth==4.8.0
flask-marshmallow==1.2.1
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
itsdangerous==2.2.0
Jinja2==3.1.4
Mako==1.3.6
MarkupSafe==3.0.2
marshmallow==3.23.0
packaging==24.1
psycopg2-binary==2.9.10
python-dotenv==1.0.1
SQLAlchemy==2.0.36
typing_extensions==4.12.2
webargs==8.6.0
Werkzeug==3.0.4
